Update Truffle Tests to Support StringValue
