+++ test/EventStore/Mock.js
@@ -1,0 +1,14 @@
+const testAddressValueEvent = {
+    Type: 'ADDRESS_EVENT_TESTED',
+    Version: 'v0',
+    ValueType: 'Address',
+    AddressValue: '0x0000000000000000000000000000000000000000',
+    UIntValue: 0,
+    Bytes32Value: '',
+    StringValue: '',
+    PropertyCount: 0
+}
+
+module.exports = {
+    testAddressValueEvent
+}
